requirements.txt
pandas
tabula-py
jpype1
jpype